+++ fabric/src/main/java/com/github/smallinger/copperagebackport/client/CopperAgeBackportFabricClient.java
@@ -105,7 +105,16 @@ public class CopperAgeBackportFabricClient implements ClientModInitializer {
             ModBlocks.WAXED_COPPER_BARS.get(),
             ModBlocks.WAXED_EXPOSED_COPPER_BARS.get(),
             ModBlocks.WAXED_WEATHERED_COPPER_BARS.get(),
-            ModBlocks.WAXED_OXIDIZED_COPPER_BARS.get());
+            ModBlocks.WAXED_OXIDIZED_COPPER_BARS.get(),
+            // Copper Trapdoors
+            ModBlocks.COPPER_TRAPDOOR.get(),
+            ModBlocks.EXPOSED_COPPER_TRAPDOOR.get(),
+            ModBlocks.WEATHERED_COPPER_TRAPDOOR.get(),
+            ModBlocks.OXIDIZED_COPPER_TRAPDOOR.get(),
+            ModBlocks.WAXED_COPPER_TRAPDOOR.get(),
+            ModBlocks.WAXED_EXPOSED_COPPER_TRAPDOOR.get(),
+            ModBlocks.WAXED_WEATHERED_COPPER_TRAPDOOR.get(),
+            ModBlocks.WAXED_OXIDIZED_COPPER_TRAPDOOR.get());
     }
 }